+++ test/resources/checkers/c_c++-clang-included-file-error.cpp
@@ -1,0 +1,3 @@
+class test {};
+
+#include <c_c++-clang-bad-header.h>