+++ test/resources/checkers/c_c++-include/c_c++-clang-bad-header.h
@@ -1,0 +1,3 @@
+class {
+  this_is_bad
+}